requirements.txt
six
numpy>=1.15
lalsuite